requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0

# HuggingFace ecosystem
transformers>=4.40.0
datasets>=2.14.0
accelerate>=0.20.0

# Vision models
unsloth
timm>=0.9.0
Pillow>=10.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Experiment tracking
wandb>=0.15.0

# Development tools
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# OCR specific
pytesseract>=0.3.10
easyocr>=1.7.0

# Knowledge distillation
scikit-learn>=1.3.0
